lib/reg_script: Add cast to fix compilation on x86_64
Reviewed-on: https://review.coreboot.org/c/coreboot/+/48166
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/lib/reg_script.c
@@ -150,11 +150,11 @@ static uint32_t reg_script_read_mmio(struct reg_script_context *ctx)
 
 	switch (step->size) {
 	case REG_SCRIPT_SIZE_8:
-		return read8((u8 *)step->reg);
+		return read8((u8 *)(uintptr_t)step->reg);
 	case REG_SCRIPT_SIZE_16:
-		return read16((u16 *)step->reg);
+		return read16((u16 *)(uintptr_t)step->reg);
 	case REG_SCRIPT_SIZE_32:
-		return read32((u32 *)step->reg);
+		return read32((u32 *)(uintptr_t)step->reg);
 	}
 	return 0;
 }
@@ -165,13 +165,13 @@ static void reg_script_write_mmio(struct reg_script_context *ctx)
 
 	switch (step->size) {
 	case REG_SCRIPT_SIZE_8:
-		write8((u8 *)step->reg, step->value);
+		write8((u8 *)(uintptr_t)step->reg, step->value);
 		break;
 	case REG_SCRIPT_SIZE_16:
-		write16((u16 *)step->reg, step->value);
+		write16((u16 *)(uintptr_t)step->reg, step->value);
 		break;
 	case REG_SCRIPT_SIZE_32:
-		write32((u32 *)step->reg, step->value);
+		write32((u32 *)(uintptr_t)step->reg, step->value);
 		break;
 	}
 }
